+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- TypeScript definition for `ecfToEci` function is added. 
+
 ## 4.1.3 (2020-12-18)
 
 - Fixed calculation of Doppler effect (`dopplerFactor`).